broke up code in index.js and fixed import/export
